requirements.txt
gunicorn==20.0.4
Flask==1.0.2
Flask-Cors==3.0.10
Flask-Session==0.3.2
Flask-SocketIO==5.0.1
eventlet==0.31.0
python-dotenv==0.16.0
Pyrebase4==4.4.3
redis==3.5.3
